[BOOK-179] refactor: apis - KakaoApiHelper를 KakaoApiManager로 리네이밍 및 예외 처리 구체화
