--- apis/src/main/kotlin/org/yapp/apis/auth/helper/KakaoApiHelper.kt
@@ -1,29 +1,0 @@
-package org.yapp.apis.auth.helper
-
-import mu.KotlinLogging
-import org.yapp.apis.auth.exception.AuthErrorCode
-import org.yapp.apis.auth.exception.AuthException
-import org.yapp.globalutils.annotation.Helper
-import org.yapp.infra.external.oauth.kakao.KakaoApi
-import org.yapp.infra.external.oauth.kakao.response.KakaoUserInfo
-
-@Helper
-class KakaoApiHelper(
-    private val kakaoApi: KakaoApi
-) {
-    private val log = KotlinLogging.logger {}
-
-    fun getUserInfo(accessToken: String): KakaoUserInfo {
-        return kakaoApi.fetchUserInfo(accessToken)
-            .onSuccess { userInfo ->
-                log.info("Successfully fetched Kakao user info for userId: ${userInfo.id}")
-            }
-            .getOrElse { exception ->
-                log.error("Failed to fetch Kakao user info", exception)
-                throw AuthException(
-                    AuthErrorCode.FAILED_TO_GET_USER_INFO,
-                    "Failed to call Kakao API: ${exception.message}"
-                )
-            }
-    }
-}
